fix: keep pkg/api at 15 files after in-flight test split
Merge session upsert tests into server_factory_test.go so adding
server_work_query_inflight_test.go does not violate pkg-file-count.
